lints, and partially working error thingy

+++ system-configuration/src/status_temp.rs
@@ -89,7 +89,7 @@ mod status_type_impls {
 
     impl From<SCStatusType> for OSStatus {
         fn from(value: SCStatusType) -> Self {
-            Into::<u8>::into(value).into()
+            Into::<u32>::into(value).into()
         }
     }
 }
